Issue 41: Step 9 (Performance)
Methods which do not access instance data or call instance 
methods can be marked as static (Shared in Visual Basic). 
After doing so, the compiler will emit non-virtual 
call sites to these members which will prevent a check 
at runtime for each call that ensures the current object 
pointer is non-null. This can result in a measurable 
performance gain for performance-sensitive code. In 
some cases, the failure to access the current object 
instance represents a correctness issue.

http://msdn.microsoft.com/tr-tr/library/ms245046(en-us).aspx


git-svn-id: https://msnp-sharp.googlecode.com/svn/trunk/MSNP-Sharp15@483 ed960e22-f52a-0410-8d30-b9dff305f3af